get_dot_dependency_graph(): Don't use repr() to get statement labels
Sometimes, it is useful to have labels that include escape sequences
significant to the dot language. The string that comes from repr()
doesn't allow us to do that because it escapes all the backslashes.

>>> print(str("a\lb"))
a\lb

>>> print(repr("a\lb"))
'a\\lb'

This change removes the use of repr() and documents that the
statements are turned into strings using str().